Count reads with featureCounts

Count reads with featureCounts

fcount(
  sample_info = NULL,
  mappingdir = "results/04_read_mapping",
  gff_path = NULL,
  fcountsdir = "results/05_quantification/featureCounts",
  threads = 2
)

Arguments

sample_info

Data frame of sample metadata created with the functions create_sample_info and infer_strandedness.

mappingdir

Directory where .bam files are stored.

gff_path

Path to GFF/GTF file with annotations.

fcountsdir

Directory where the matrix of gene-level read counts will be stored.

threads

Number of threads for featureCounts. Default: 2.

Value

A gene expression matrix with genes in row names and samples in column names. Two .tsv files with the gene expression matrix and count stats are saved to fcountsdir.

Examples

data(sample_info)
mappingdir <- system.file("extdata", package="bears")
gff_path <- system.file("extdata", "Homo_sapiens.GRCh37.75_subset.gtf",
                        package="bears")
fcountsdir <- tempdir()
if(subread_is_installed()) {
    counts <- fcount(sample_info, mappingdir, gff_path, fcountsdir)
}