Create a SummarizedExperiment object from kallisto output

Create a SummarizedExperiment object from kallisto output

kallisto2se(
  sample_info = NULL,
  level = "gene",
  kallistodir = "results/05_quantification/kallisto",
  tx2gene = NULL
)

Arguments

sample_info

Data frame of sample metadata created with the functions create_sample_info

level

Character indicating to which level expression must be quantified in the SE object. One of "gene" (default), "transcript", or "both". For "both", the SE object will have two assays named "transcript" and "gene".

kallistodir

Directory where quantification files will be stored. Default: results/05_quantification/kallisto.

tx2gene

Data frame of correspondence between genes and transcripts, with gene IDs in the first column and transcript IDs in the second column. Only required if level = 'gene' or 'both'.

Value

A SummarizedExperiment object with gene/transcript expression levels and sample metadata.

Examples

data(sample_info)
data(tx2gene)
kallistodir <- system.file("extdata", package="bears")
se_gene <- kallisto2se(sample_info, kallistodir = kallistodir, 
                       tx2gene = tx2gene)
#> Note: importing `abundance.h5` is typically faster than `abundance.tsv`
#> reading in files with read.delim (install 'readr' package for speed up)
#> 1 
#> 
#> summarizing abundance
#> summarizing counts
#> summarizing length