Quantify expression with salmon

Quantify expression with salmon

salmon_quantify(
  sample_info = NULL,
  filtdir = "results/03_filtered_FASTQ",
  salmonindex = "results/05_quantification/salmon/idx",
  salmondir = "results/05_quantification/salmon",
  threads = NULL
)

Arguments

sample_info

Data frame of sample metadata created with the functions create_sample_info and infer_strandedness. The function infer_strandedness adds a column named "Orientation" with library strandedness information. If this column is not present in sample_info, salmon will automatically infer the library strandedness, but it will take longer to run.

filtdir

Path to the directory where filtered reads are stored. Default: results/03_filtered_FASTQ.

salmonindex

Directory where the transcriptome index is stored. Default: results/05_quantification/salmon/idx.

salmondir

Directory where quantification files will be stored. Default: results/05_quantification/salmon.

threads

Number of threads for salmon quant.

Value

A 2-column data frame with BioSample IDs in the first column and salmon quantification status in the second column, with "OK" if salmon sucessfully quantified expression for a given BioSample, and NA otherwise.

Examples

data(sample_info)
filtdir <- system.file("extdata", package = "bears")
salmonindex <- tempdir()
salmondir <- tempdir()
transcriptome_path <- system.file(
     "extdata", "Hsapiens_GRCh37.75_subset_transcripts.fa", package="bears"
)
if(salmon_is_installed()) {
    salmon_index(salmonindex, transcriptome_path)
    salmon_quantify(sample_info, filtdir, salmonindex, salmondir)
}